restrucutre advent common
