+++ advent-common/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2024"
 
 [dependencies]
 paste = { workspace = true }
+thiserror = { workspace = true }
