+++ solutions/Cargo.toml
@@ -8,4 +8,3 @@ advent-common = { path = "../advent-common" }
 clap = { workspace = true }
 paste = { workspace = true }
 regex = { workspace = true }
-thiserror = { workspace = true }
