+++ solutions/src/lib.rs
@@ -1,38 +1,7 @@
 use std::borrow::Cow;
-use std::fs::File;
-use std::io::{self, BufRead};
-use std::num::ParseIntError;
 use std::path::PathBuf;
-use thiserror::Error;
 
-pub trait AdventProblem {
-    fn run_part_1(&self, lines: Vec<String>) -> Result<u32, AdventError>;
-    fn run_part_2(&self, lines: Vec<String>) -> Result<u32, AdventError>;
-
-    fn run(&self, lines: Vec<String>, part: u8) -> Result<u32, AdventError> {
-        match part {
-            1 => self.run_part_1(lines),
-            2 => self.run_part_2(lines),
-            _ => unimplemented!("part {}", part),
-        }
-    }
-}
-
-#[derive(Error, Debug)]
-pub enum AdventError {
-    #[error("{0}")]
-    IoError(#[from] io::Error),
-    #[error("{0}")]
-    InputParseError(Cow<'static, str>),
-    #[error("unknown error")]
-    UnknownError,
-}
-
-impl From<ParseIntError> for AdventError {
-    fn from(err: ParseIntError) -> Self {
-        Self::InputParseError(err.to_string().into())
-    }
-}
+pub use advent_common::{AdventError, AdventProblem};
 
 #[macro_export]
 macro_rules! regex {
@@ -48,21 +17,12 @@ pub mod y2023;
 pub mod y2024;
 pub mod y2025;
 
-fn read_input_lines(year: u32, date: u8) -> Result<Vec<String>, AdventError> {
-    let path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .join("inputs")
-        .join(format!("y{}", year))
-        .join(format!("day{}.txt", date));
-    let mut lines = Vec::new();
-    let file = File::open(path)?;
-    for next in io::BufReader::new(file).lines() {
-        lines.push(next?);
-    }
-    Ok(lines)
-}
-
 pub fn run(year: u32, date: u8, part: u8) -> Result<u32, AdventError> {
-    let lines = read_input_lines(year, date)?;
+    let lines = advent_common::read_input_lines(
+        PathBuf::from(env!("CARGO_MANIFEST_DIR")),
+        year,
+        date,
+    )?;
     match year {
         2023 => y2023::run_with_lines(lines, date, part),
         2024 => y2024::run_with_lines(lines, date, part),
