+++ solutions/src/y2024/mod.rs
@@ -1,4 +1,4 @@
-use super::{AdventError, AdventProblem};
+use super::AdventError;
 
 advent_common::define_advent_registry!(
     1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
